Update README.md
This commit marks an important waypoint in my PhD: I decided to see it
through, and this is the start of the final push.

If I succeed, it will be dedicated to all the failures of global
scientific establishment in late-stage capitalism and to the intrinsic
hypocrisy of this system. Fuck your sanctions against scientific
institutions, fuck your regional bans. The cake is a lie.