Update crop_by_other to work on skewed point sets
Most of the ground plots are rotated rectangles, and simple crop by
extent produced crops that included too many extra points. The new
approach uses scipy.spatial to construct a Delaunay triangulation for
the vertex points of the convex hull of `other` and check if the
`points` are within any of the triangles.