Add a function to create a grid from  clouds
This is the initial version that I used as a proof of concept in a
Kaggle notebook. It operates on Numpy arrays of coordinates, and returns
the combined coordinate array and a label array. It should probably
operate on laspy.LasData, and also return the combined array of features
(i.e., color).