Update the DataSphere job config
It still doesn't launch and doesn't generate any logs, which is annoying
beyond comprehension. I am choosing to give up with this for now since
there is an easier way to run on DataSphere – through JupyterLab. Maybe
one day, when there is no hurry, I will come back and figure this out.